Reorganized scripts a little (#7902)
